+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_metrics_query_client.py
@@ -53,7 +53,7 @@ class MetricsQueryClient(object):
         self._namespace_op = self._client.metric_namespaces
         self._definitions_op = self._client.metric_definitions
 
-    def query(self, resource_uri, metric_names, duration=None, **kwargs):
+    def query(self, resource_uri, metric_names, **kwargs):
         # type: (str, list, Optional[timedelta], Any) -> MetricsResult
         """Lists the metric values for a resource.
 
@@ -64,13 +64,10 @@ class MetricsQueryClient(object):
         :type resource_uri: str
         :param metric_names: The names of the metrics to retrieve.
         :type metric_names: list[str]
-        :param ~datetime.timedelta duration: The duration for which to query the data. This can also be accompanied
-         with either start_time or end_time. If start_time or end_time is not provided, the current time is
-         taken as the end time.
-        :keyword datetime start_time: The start time from which to query the data. This should be accompanied
-         with either end_time or duration.
-        :keyword datetime end_time: The end time till which to query the data. This should be accompanied
-         with either start_time or duration.
+        :keyword timespan: The timespan for which to query the data. This can be a timedelta,
+         a timedelta and a start datetime, or a start datetime/end datetime.
+        :paramtype timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
+         or tuple[~datetime.datetime, ~datetime.datetime]
         :keyword interval: The interval (i.e. timegrain) of the query.
         :paramtype interval: ~datetime.timedelta
         :keyword aggregations: The list of aggregation types to retrieve. Use `azure.monitor.query.AggregationType`
@@ -112,12 +109,11 @@ class MetricsQueryClient(object):
             :dedent: 0
             :caption: Get a response for a single Metrics Query
         """
-        start = kwargs.pop('start_time', None)
-        end = kwargs.pop('end_time', None)
+
         aggregations = kwargs.pop("aggregations", None)
         if aggregations:
             kwargs.setdefault("aggregation", ",".join(aggregations))
-        timespan = construct_iso8601(start, end, duration)
+        timespan = construct_iso8601(kwargs.pop("timespan", None))
         kwargs.setdefault("metricnames", ",".join(metric_names))
         kwargs.setdefault("timespan", timespan)
         generated = self._metrics_op.list(resource_uri, connection_verify=False, **kwargs)