+++ sdk/monitor/azure-monitor-query/samples/async_samples/sample_log_query_client_async.py
@@ -24,7 +24,7 @@ async def logs_query():
 
     # returns LogsQueryResult
     async with client:
-        response = await client.query(os.environ['LOG_WORKSPACE_ID'], query)
+        response = await client.query(os.environ['LOG_WORKSPACE_ID'], query, timespan=None)
 
     if not response.tables:
         print("No results for the query")